--- src/components/SignUpForm/ErrorUser/ErrorUserModal.styled.js
@@ -1,168 +1,0 @@
-import styled from 'styled-components';
-import { Form, Field } from 'formik';
-import errorIcon from '../../../assets/images/icons/error.svg';
-import correctIcon from '../../../assets/images/icons/correct.svg';
-import eyeOpen from '../../../assets/images/icons/eyeOpen.svg';
-import eyeClose from '../../../assets/images/icons/eyeClose.svg';
-
-export const Backdrop = styled.div`
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  background-color: rgba(5, 5, 5, 0.8);
-`;
-
-export const WaterIntakeModal = styled.div`
-  position: absolute;
-  top: 60px;
-  left: 50%;
-  transform: translateX(-50%);
-  z-index: 999;
-
-  width: 300px;
-  height: 266px;
-  border-radius: 12px;
-  padding: 16px 12px;
-  background-color: var(--bg-secondary);
-
-  font-family: Poppins;
-  font-style: normal;
-
-  @media screen and (min-width: 834px) {
-    top: 180px;
-    width: 338px;
-    height: 280px;
-    padding: 24px 40px;
-  }
-
-  @media screen and (min-width: 1440px) {
-    top: 215px;
-  }
-`;
-
-export const WaterIntakeForm = styled(Form)`
-  position: relative;
-  display: flex;
-  flex-direction: column;
-
-  @media screen and (min-width: 834px) {
-    width: 258px;
-    padding-left: 23px;
-    padding-right: 23px;
-  }
-`;
-
-export const Title = styled.h3`
-  font-size: 24px;
-  font-weight: 500;
-  line-height: 30px;
-  margin-bottom: 24px;
-
-  @media screen and (min-width: 834px) {
-    font-size: 30px;
-    line-height: 36px;
-  }
-`;
-
-export const WaterLable = styled.label`
-  position: relative;
-  font-size: 14px;
-  font-weight: 500;
-  line-height: 20px;
-  margin-bottom: 12px;
-
-  &::after {
-    position: absolute;
-    bottom: -38px;
-    right: 10px;
-    content: '';
-    display: ${({ $showIcon }) => $showIcon};
-    margin-left: 16px;
-    width: 16px;
-    height: 16px;
-    background-image: url('${errorIcon}');
-  }
-`;
-
-export const InputField = styled(Field)`
-  width: 276px;
-  height: 36px;
-  padding: 8px 10px;
-  background-color: var(--bg-secondary);
-  border: 1px solid var(--input-border-color-normal);
-  border-radius: 12px;
-  margin-bottom: 16px;
-  outline: none;
-
-  &::-webkit-inner-spin-button {
-    -webkit-appearance: none;
-  }
-
-  color: var(--text-color-secondary-grey);
-  font-family: Poppins;
-  font-size: 14px;
-  font-weight: 400;
-  line-height: 20px;
-  border: ${({ borderstyle }) => borderstyle};
-
-  @media screen and (min-width: 834px) {
-    width: 212px;
-  }
-`;
-
-export const ConfirmButton = styled.button`
-  width: 276px;
-  height: 36px;
-  padding: 8px 10px;
-  border-radius: 12px;
-  margin-bottom: 8px;
-  background-color: var(--bg-button-color);
-  border: none;
-
-  color: var(--text-color-button-black);
-  font-family: Poppins;
-  font-size: 14px;
-  font-weight: 500;
-  line-height: 20px;
-
-  @media screen and (min-width: 834px) {
-    width: 212px;
-  }
-`;
-
-export const ErrorText = styled.div`
-  position: absolute;
-  top: 69px;
-  left: 0;
-  color: #e74a3b;
-  font-family: Poppins;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 14px;
-  padding-left: 10px;
-
-  @media screen and (min-width: 834px) {
-    left: 25px;
-  }
-`;
-
-export const CancelButton = styled.button`
-  width: 276px;
-  height: 36px;
-  padding: 8px 10px;
-  border-radius: 12px;
-  background-color: transparent;
-  border: none;
-
-  color: var(--text-color-secondary-grey);
-  font-family: Poppins;
-  font-size: 14px;
-  line-height: 20px;
-
-  @media screen and (min-width: 834px) {
-    width: 212px;
-  }
-`;
