+++ src/components/SignUpForm/ErrorUserModal/ErrorUserModal.styled.js
@@ -1,0 +1,73 @@
+import styled from 'styled-components';
+
+// import errorIcon from '../../../assets/images/icons/error.svg';
+// import correctIcon from '../../../assets/images/icons/correct.svg';
+// import eyeOpen from '../../../assets/images/icons/eyeOpen.svg';
+// import eyeClose from '../../../assets/images/icons/eyeClose.svg';
+
+export const Backdrop = styled.div`
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+
+  background-color: rgba(5, 5, 5, 0.8);
+`;
+
+export const Modal = styled.div`
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translateX(-50%);
+  z-index: 999;
+
+  width: 300px;
+  height: 266px;
+  border-radius: 12px;
+  padding: 16px 12px;
+  background-color: var(--bg-secondary);
+
+  font-family: Poppins;
+  font-style: normal;
+
+  @media screen and (min-width: 834px) {
+    top: 180px;
+    width: 338px;
+    height: 280px;
+    padding: 24px 40px;
+  }
+
+  @media screen and (min-width: 1440px) {
+    top: 215px;
+  }
+`;
+
+export const Title = styled.h3`
+  font-size: 24px;
+  font-weight: 500;
+  line-height: 30px;
+  margin-bottom: 24px;
+
+  @media screen and (min-width: 834px) {
+    font-size: 30px;
+    line-height: 36px;
+  }
+`;
+
+export const ErrorText = styled.div`
+  position: absolute;
+  top: 69px;
+  left: 0;
+  color: #e74a3b;
+  font-family: Poppins;
+  font-size: 12px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 14px;
+  padding-left: 10px;
+
+  @media screen and (min-width: 834px) {
+    left: 25px;
+  }
+`;
